+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add ability to push image to registry with `Image::push` or `Images::push`.
 - Add `online_cpus` field to `CpuStats`
 - Fix `Image::history` response deserialization
+- Fix `Container::logs` and `Service::logs` endpoints
 
 # 0.6.0
 - `name` field of `ContainerCreateOpts` is now private. Use the `ContainerCreateOpts::builder` function that takes in a `name` parameter.